enable instruction parsing

+++ SimWorld/OpenGL/LTree/TreeBranch.h
@@ -49,7 +49,7 @@
                andLength:(float)length
                 andStart:(float)startRadius
                   andEnd:(float)endRadius
-          andParentIndex:(int)parentIndex
-       andParentPosition:(int)parentPosition;
+          andParentIndex:(NSInteger)parentIndex
+       andParentPosition:(float)parentPosition;
 
 @end
